+++ src/database/postgres.rs
@@ -16,12 +16,12 @@ impl Postgres {
         }
     }
 
-    fn run_counting_query(&self, query: &str, output_column_name: &str) -> usize {
+    fn run_counting_query(&self, query: &str, output_column_name: &str) -> u32 {
         if let Some(mut client) = self.get_client() {
             if let Ok(rows) = client.query(&*query, &[]) {
                 if let Some(row) = rows.get(0) {
                     let sum: i64 = row.get(output_column_name);
-                    return sum as usize;
+                    return sum as u32;
                 }
             }
         }
@@ -56,7 +56,7 @@ impl DatabaseInterface for Postgres {
         }
     }
 
-    fn get_count_of_all_queries_for_table(&self, table_name: &str) -> usize {
+    fn get_count_of_all_queries_for_table(&self, table_name: &str) -> u32 {
         let query = format!(
             "SELECT SUM(calls::INTEGER) FROM pg_stat_statements WHERE query ~* '[[:<:]]{}[[:>:]]'",
             table_name
@@ -65,13 +65,13 @@ impl DatabaseInterface for Postgres {
         self.run_counting_query(&query, "sum")
     }
 
-    fn get_count_of_rows_selected_for_table(&self, table_name: &str) -> usize {
+    fn get_count_of_rows_selected_for_table(&self, table_name: &str) -> u32 {
         let query = format!("SELECT SUM(rows::INTEGER) FROM pg_stat_statements WHERE query ~* '[[:<:]]{}[[:>:]]' AND query ~* 'select'", table_name);
 
         self.run_counting_query(&query, "sum")
     }
 
-    fn get_count_of_rows_updated_for_table(&self, table_name: &str) -> usize {
+    fn get_count_of_rows_updated_for_table(&self, table_name: &str) -> u32 {
         let query = format!("SELECT SUM(rows::INTEGER) FROM pg_stat_statements WHERE query ~* '[[:<:]]{}[[:>:]]' AND query ~* 'update'", table_name);
 
         self.run_counting_query(&query, "sum")
