+++ src/test_type/json.rs
@@ -7,7 +7,7 @@ use serde_json::Value;
 use std::cmp::min;
 
 pub struct Json {
-    pub concurrency_levels: Vec<usize>,
+    pub concurrency_levels: Vec<u32>,
 }
 impl Executor for Json {
     fn retrieve_benchmark_commands(&self, url: &str) -> VerifierResult<BenchmarkCommands> {
@@ -41,7 +41,7 @@ impl Executor for Json {
     }
 }
 impl Json {
-    fn get_wrk_command(&self, url: &str, duration: usize, concurrency: usize) -> Vec<String> {
+    fn get_wrk_command(&self, url: &str, duration: u32, concurrency: u32) -> Vec<String> {
         vec![
             "wrk",
             "-H",
@@ -58,7 +58,7 @@ impl Json {
             "--timeout",
             "8",
             "-t",
-            &format!("{}", min(concurrency, num_cpus::get())),
+            &format!("{}", min(concurrency, num_cpus::get() as u32)),
             url,
         ].iter().map(|item| item.to_string()).collect()
     }
