+++ src/test_type/plaintext.rs
@@ -6,7 +6,7 @@ use crate::verification::Messages;
 use std::cmp::min;
 
 pub struct Plaintext {
-    pub pipeline_concurrency_levels: Vec<usize>,
+    pub pipeline_concurrency_levels: Vec<u32>,
 }
 impl Executor for Plaintext {
     fn retrieve_benchmark_commands(&self, url: &str) -> VerifierResult<BenchmarkCommands> {
@@ -68,7 +68,7 @@ impl Plaintext {
         }
     }
 
-    fn get_wrk_command(&self, url: &str, duration: usize, concurrency: usize) -> Vec<String> {
+    fn get_wrk_command(&self, url: &str, duration: u32, concurrency: u32) -> Vec<String> {
         vec![
             "wrk",
             "-H",
@@ -85,7 +85,7 @@ impl Plaintext {
             "--timeout",
             "8",
             "-t",
-            &format!("{}", min(concurrency, num_cpus::get())),
+            &format!("{}", min(concurrency, num_cpus::get() as u32)),
             url,
         ].iter().map(|item| item.to_string()).collect()
     }
