+++ src/test_type/query/updates.rs
@@ -9,7 +9,7 @@ use std::cmp;
 use std::cmp::min;
 
 pub struct Updates {
-    pub concurrency_levels: Vec<usize>,
+    pub concurrency_levels: Vec<u32>,
     pub database_verifier: Box<dyn DatabaseInterface>,
 }
 impl Query for Updates {}
@@ -107,9 +107,9 @@ impl Updates {
         &self,
         url: &str,
         table_name: &str,
-        concurrency: usize,
-        repetitions: usize,
-        expected_updates: usize,
+        concurrency: u32,
+        repetitions: u32,
+        expected_updates: u32,
         messages: &mut Messages,
     ) {
         let all_rows_updated_before_count = self
@@ -145,8 +145,8 @@ impl Updates {
     fn verify_updates(
         &self,
         url: &str,
-        concurrency: usize,
-        repetitions: usize,
+        concurrency: u32,
+        repetitions: u32,
         messages: &mut Messages,
     ) {
         let expected_updates = concurrency * repetitions;
@@ -188,7 +188,7 @@ impl Updates {
         }
     }
 
-    fn get_wrk_command(&self, url: &str, duration: usize, concurrency: usize) -> Vec<String> {
+    fn get_wrk_command(&self, url: &str, duration: u32, concurrency: u32) -> Vec<String> {
         vec![
             "wrk",
             "-H",
@@ -205,7 +205,7 @@ impl Updates {
             "--timeout",
             "8",
             "-t",
-            &format!("{}", min(concurrency, num_cpus::get())),
+            &format!("{}", min(concurrency, num_cpus::get() as u32)),
             url,
         ].iter().map(|item| item.to_string()).collect()
     }
